+++ project/checks.c
@@ -3,10 +3,10 @@
 #include "LBDefinitions.h"
 #include "checks.h"
 
-void check_in_rank(double *collideField, int *flagField, int * length, int t){
+void check_in_rank(float *collideField, int *flagField, int * length, int t){
 	int node[D], i, x, y, z;
 	int n[D] = { length[0] + 2, length[1] + 2, length[2] + 2 };
-	double velocity[D], density, norm_v=0, *currentCell;
+	float velocity[D], density, norm_v=0, *currentCell;
 
 	for (z = 1; z < length[2]; ++z){
 		node[2] = z;
@@ -77,7 +77,7 @@ void check_mass(double *massField, int* flagField, int* length, int t){
 	printf("On timestep %d : Total mass: %f \n",t,tot_mass);
 }
 
-void run_checks(double *collideField, double *massField, int *flagField, int * length, int t ){
+void run_checks(float *collideField, double *massField, int *flagField, int * length, int t ){
 //	check_in_rank(collideField, flagField, length, t);
 //	check_flags(flagField, length, FLUID, GAS, 	 t);
 //	check_mass( massField, flagField, length, t);
