+++ project/computeCellValues.h
@@ -5,15 +5,15 @@
  *  currentCell thus denotes the address of the first particle distribution function of the
  *  respective cell. The result is stored in density.
  */
-void computeDensity(const double *const currentCell, double *density);
+void computeDensity(const float *const currentCell, float *density);
 
 /** computes the velocity within currentCell and stores the result in velocity */
-void computeVelocity(const double *const currentCell, const double * const density,double *velocity);
+void computeVelocity(const float *const currentCell, const float * const density,float *velocity);
 
 /** computes the equilibrium distributions for all particle distribution functions of one
  *  cell from density and velocity and stores the results in feq.
  */
-void computeFeq(const double * const density, const double * const velocity, double *feq);
+void computeFeq(const float * const density, const float * const velocity, float *feq);
 
 #endif
