+++ project/flag.h
@@ -1,6 +1,6 @@
 #ifndef _FLAG_H_
 #define _FLAG_H_
 
-void updateFlagField(double * collideField, int * flagField, double * fractionField, int * length);
+void updateFlagField(float * collideField, int * flagField, double * fractionField, int * length);
 
 #endif
